Large update/rewrite to support magnetic fields
The B field points in the z-direction, the electric field lies in the
y,z plane. The angle between these two can be specified.

An analytic particle mover is included, as well as the Boris method,
similarly implemented as was done by Anbang Sun.

The output format for the transport data was changed, as well as the
accuracy testing. The mobility and diffusion coefficient are now given
per coordinate (x,y,z)